requirements.txt
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
colorama==0.4.6
filelock==3.29.0
fsspec==2026.3.0
idna==3.13
Jinja2==3.1.6
Levenshtein==0.27.3
llvmlite==0.47.0
MarkupSafe==3.0.3
more-itertools==11.0.2
mpmath==1.3.0
networkx==3.6.1
numba==0.65.1
numpy==2.4.4
openai-whisper==20250625
packaging==26.0
pandas==3.0.2
pycparser==3.0
pypinyin==0.55.0
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.3
RapidFuzz==3.14.5
regex==2026.4.4
requests==2.33.1
scipy==1.17.1
setuptools==81.0.0
six==1.17.0
sounddevice==0.5.5
sympy==1.14.0
tiktoken==0.12.0
torch==2.11.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
wheel==0.46.3
